refactor: cache release data
