+++ apps/site/layouts/Download.tsx
@@ -4,11 +4,14 @@ import { getClientContext } from '#site/client-context';
 import WithDownloadSection from '#site/components/withDownloadSection';
 import WithFooter from '#site/components/withFooter';
 import WithNavBar from '#site/components/withNavBar';
+import provideReleaseData from '#site/next-data/providers/releaseData';
 
 import styles from './layouts.module.css';
 
-const DownloadLayout: FC<PropsWithChildren> = ({ children }) => {
+const DownloadLayout: FC<PropsWithChildren> = async ({ children }) => {
   const { frontmatter } = getClientContext();
+
+  const releases = await provideReleaseData();
 
   return (
     <>
@@ -18,7 +21,9 @@ const DownloadLayout: FC<PropsWithChildren> = ({ children }) => {
         <main>
           <h1>{frontmatter.title}</h1>
 
-          <WithDownloadSection>{children}</WithDownloadSection>
+          <WithDownloadSection releases={releases}>
+            {children}
+          </WithDownloadSection>
         </main>
       </div>
